added handlers and working on login

+++ src/result.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize)]
 pub struct MResult<T> {
-    pub status: i32,
+    pub status: u16,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub description: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
